# openapi.yaml
openapi: 3.1.0
info:
  title: Kitsu DeFi API
  description: |
    Gamified DeFi savings app on TON blockchain.
    
    ## Authentication
    Most endpoints require JWT authentication. Get a token by connecting your wallet via `/auth/connect`.
    
    Include the token in the `Authorization` header as `Bearer <token>`.
  version: 1.0.0
  contact:
    name: Kitsu DeFi
    url: https://kitsu.fi
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.kitsu.fi
    description: Production
...
paths:
...
  /goals:
    get:
      tags: [Goals]
      summary: List goals
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user's goals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalsResponse'
    post:
      tags: [Goals]
      summary: Create goal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Goal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
  /goals/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Goals]
      summary: Update goal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalRequest'
      responses:
        '200':
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
    delete:
      tags: [Goals]
      summary: Archive goal
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Goal archived
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      archived:
                        type: boolean
...
  /quests:
    get:
      tags: [Quests]
      summary: List quests
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's quests with progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestsResponse'
  /quests/{id}/claim:
    post:
      tags: [Quests]
      summary: Claim quest reward
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Quest claimed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimQuestResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/connect
  schemas:
...
    Success:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
...
    Goal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        emoji:
          type: string
          nullable: true
        targetUsd:
          type: string
          description: Target amount in USD
        currentUsd:
          type: string
          description: Current saved amount in USD
        dueDate:
          type: string
          format: date-time
          nullable: true
        isArchived:
          type: boolean
        createdAt:
          type: string
          format: date-time
    GoalResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                goal:
                  $ref: '#/components/schemas/Goal'
    GoalsResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                goals:
                  type: array
                  items:
                    $ref: '#/components/schemas/Goal'
    CreateGoalRequest:
      type: object
      required: [title, targetUsd]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 256
        emoji:
          type: string
          maxLength: 8
        targetUsd:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
          description: Target amount in USD
        dueDate:
          type: string
          format: date-time
    UpdateGoalRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 256
        emoji:
          type: string
          maxLength: 8
        targetUsd:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
        dueDate:
          type: string
          format: date-time
          nullable: true

    # Den
...
    Quest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        questKey:
          type: string
        title:
          type: string
        hint:
          type: string
        reward:
          type: integer
          description: XP reward
        progress:
          type: integer
        completed:
          type: boolean
    QuestsResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                quests:
                  type: array
                  items:
                    $ref: '#/components/schemas/Quest'
    ClaimQuestResponse:
      allOf:
        - $ref: '#/components/schemas/Success'
        - type: object
          properties:
            data:
              type: object
              properties:
                claimed:
                  type: boolean
                xpAwarded:
                  type: integer
                newTotalXp:
                  type: integer

    # AI Chat
...